// src/pages/Contest/ContestSider/ParticipantActive/ContestPending/types.ts
export interface ContestPendingProps {}
export interface ContestPageSiderParticipantActiveContestPendingInterface {
    readonly isFetching: boolean;
    readonly isRefreshing: boolean;
    readonly isInitialized: boolean;
    readonly err?: Error;
    readonly data: any;
    readonly metadata: any;
    readonly append: boolean;
}